Pulling retention enforcement variables out to config

+++ cv-data-tasks/src/main/java/com/trihydro/tasks/actions/RetentionPolicyEnforcement.java
@@ -5,6 +5,7 @@ import com.trihydro.library.service.DriverAlertService;
 import com.trihydro.library.service.HmiLogService;
 import com.trihydro.library.service.StatusLogService;
 import com.trihydro.library.service.TimService;
+import com.trihydro.tasks.config.DataTasksConfiguration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -12,19 +13,22 @@ import org.springframework.stereotype.Component;
 @Component
 public class RetentionPolicyEnforcement implements Runnable {
     private Utility utility;
-    DriverAlertService driverAlertService;
-    HmiLogService hmiLogService;
-    StatusLogService statusLogService;
-    TimService timService;
+    private DriverAlertService driverAlertService;
+    private HmiLogService hmiLogService;
+    private StatusLogService statusLogService;
+    private TimService timService;
+    private DataTasksConfiguration config;
 
     @Autowired
     public void InjectDependencies(Utility _utility, DriverAlertService _driverAlertService,
-            HmiLogService _hmiLogService, StatusLogService _statusLogService, TimService _timService) {
+            HmiLogService _hmiLogService, StatusLogService _statusLogService, TimService _timService,
+            DataTasksConfiguration _config) {
         this.utility = _utility;
         this.driverAlertService = _driverAlertService;
         this.hmiLogService = _hmiLogService;
         this.statusLogService = _statusLogService;
         this.timService = _timService;
+        this.config = _config;
     }
 
     public void run() {
@@ -32,14 +36,25 @@ public class RetentionPolicyEnforcement implements Runnable {
 
         try {
             // delete all older than a month:
-            // TIM
-            timService.deleteOldTim();
-            // Driver Alerts
-            driverAlertService.deleteOldDriverAlerts();
-            // HMI Logs
-            hmiLogService.deleteOldHmiLogs();
-            // Status Log
-            statusLogService.deleteOldStatusLogs();
+            if (config.getRetention_removeTims()) {
+                // TIM
+                timService.deleteOldTim();
+            }
+
+            if (config.getRetention_removeDa()) {
+                // Driver Alerts
+                driverAlertService.deleteOldDriverAlerts();
+            }
+
+            if (config.getRetention_removeHmi()) {
+                // HMI Logs
+                hmiLogService.deleteOldHmiLogs();
+            }
+
+            if (config.getRetention_removeStatusLogs()) {
+                // Status Log
+                statusLogService.deleteOldStatusLogs();
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
